refactor : BidInfo와 모달창 로직 분리
